// src/FrontEnd/src/app/models/MerchantInfo.ts
export interface MerchantInfo {
    MerchantId: string;
    MerchantCode: string;
    ServiceUse: string;
    ServiceStartDate: string;
    SuspenseReceipts: number;
    FeeType: string;
    MinFee: number;
    MinLimit: number;
    FeeByCase: number;
    MonthBaseFee: number;
    CalculateDay: number;
}